refactor(#334): 신고하기 공통 컴포넌트 분리
